+++ RPG/ArmaArqueiro.java
@@ -1,9 +1,12 @@
+//* Essa classe só poderá ser acessada pelo Arqueiro, pois apartir dela serão criados armamentos para uso do mesmo *//
+
 public class ArmaArqueiro extends Arma {
     
-    public ArmaArqueiro(String nomeArma, int atqArma, int defArma) {
+    public ArmaArqueiro(String nomeArma, double atqArma, double defArma) {
         
         super(nomeArma, atqArma, defArma);
 
     }
+
 
 }
