+++ RPG/ArmaGuerreiro.java
@@ -1,6 +1,8 @@
+//* Essa classe só poderá ser acessada pelo Guerreiro, pois apartir dela serão criados armamentos para uso do mesmo *//
+
 public class ArmaGuerreiro extends Arma {
 
-    public ArmaGuerreiro(String nomeArma, int atqArma, int defArma) {
+    public ArmaGuerreiro(String nomeArma, double atqArma, double defArma) {
         
         super(nomeArma, atqArma, defArma);
