+++ RPG/ArmaMago.java
@@ -1,6 +1,8 @@
+//* Essa classe só poderá ser acessada pelo Mago, pois apartir dela serão criados armamentos para uso do mesmo *//
+
 public class ArmaMago extends Arma {
 
-    public ArmaMago(String nomeArma, int atqArma, int defArma) {
+    public ArmaMago(String nomeArma, double atqArma, double defArma) {
         
         super(nomeArma, atqArma, defArma);
