+++ RPG/Dragao.java
@@ -1,3 +1,5 @@
+//* Personagem não jogavel, ele servirá para atacar os personagens jogaveis *//
+
 public class Dragao extends Personagem {
 
     public Dragao () {
@@ -6,9 +8,9 @@ public class Dragao extends Personagem {
 
     @Override
     public void atacar (Personagem personagem) {
-        int atqComArma;
+        double atqComArma;
         atqComArma = this.getAtqBase();
-        System.out.printf("%s: \nUsou sua bola de fogo para atacar o %s,\npontos de ataque: %d\n", 
+        System.out.printf("%s: \nUsou sua bola de fogo para atacar o %s,\npontos de ataque: %.1f\n", 
                             this.getNome(), 
                             personagem.getClasse(), 
                             atqComArma);
@@ -16,20 +18,21 @@ public class Dragao extends Personagem {
     
     @Override
     public void defenderAtaque (Personagem personagem) {
-        int dano = (personagem.getAtqBase() + personagem.getArma().getAtqArma()) - this.getDefBase();
+        double dano = (personagem.getAtqBase() + personagem.getArma().getAtqArma()) - this.getDefBase();
         if (dano < 0){
             dano = 0;
         }
         this.setVida(this.getVida() - dano);
 
-        System.out.printf("%s: \nDefendeu o ataque com suas escamas,\npontos de defesa: %d\n", 
+        System.out.printf("%s: \nDefendeu o ataque com suas escamas,\npontos de defesa: %.1f\n", 
                                 this.getNome(), 
                                 this.getDefBase());
         System.out.println();  
-        System.out.printf("DANO FINAL: %d", 
+        System.out.printf("DANO FINAL: %.1f", 
                                 dano);
-        System.out.printf("\nSTATUS DE VIDA DO %s: %d\n", 
+        System.out.printf("\nSTATUS DE VIDA DO %s: %.1f\n", 
                                 this.getNome(), 
                                 this.getVida());        
     }
+
 }
